Add current functionality of indexing and matching

+++ src/NzbDrone.Core/Mangas/Manga.cs
@@ -10,7 +10,7 @@ public class Manga : ModelBase
         Titles = new List<string>();
     }
 
-    public int? MangaUpdatesId { get; set; }
+    public long? MangaUpdatesId { get; set; }
     public int? MyAnimeListId { get; set; }
     public int? AniListId { get; set; }
     public List<string> Titles { get; set; }
